rename interfaces.ts to types.ts
