+++ src/components/PalettePanel.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Palette from "../models/Palette";
 import PaletteDisplay from "./PaletteDisplay";
 import QuantizeButton from "./buttons/QuantizeButton";
-import { Color, EditorSettings } from "../util/interfaces";
+import { Color, EditorSettings } from "../util/types";
 
 interface PalettePanelProps {
   palette: Palette;
