+++ src/util/consts.ts
@@ -1,4 +1,4 @@
-import { Dimensions, Color, EditorMode, EditorSettings } from "./interfaces";
+import { Dimensions, Color, EditorMode, EditorSettings } from "./types";
 
 export const COLORS = {
   black: new Color(0, 0, 0, 1),
